refactor(runtime): extract config materialization
Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)
